+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* `company-auto-complete` and `company-auto-complete-chars` have been renamed to
+  `company-auto-commit` and `company-auto-commit-chars` respectively.
 * `company-clang` filters out duplicates
   ([#841](https://github.com/company-mode/company-mode/issues/841)).
 * New user option `company-tooltip-width-grow-only`.